moved peopleReq
